Initial draft of Jest infrastructure for platform-specific tests with snapshots. Includes a sample test with actual snapshots.
This draft doesn't work. We use the same set of tests for each platform, but we have run Jest once per platform due to the way it is configured (haste object). Each run produces a platform-specific snapshot for the same test. The snapshot cache gets confused by this. Subsequent runs of any one platform lead to the snapshot detector saying that there are a number of obsolete snapshots. For example, running ios causes android, windows, and macos snapshots to show as obsolete.

sdx-platform has solved this somehow. The code is very large and complex, and heavily tied to react-native.